get norm from plane and get angle from vector functions
Added private functions to leg class that will be used for IK function later on. Possible errors are if I mixed up the correct orientation of the vectors in the cross product. Also, in the last commit I used sin and cos for update_coxa_to_femur we will want to discuss to make sure I used the right ones.